deploy dask configmap (#513)
* feat: creat dask configmap when jupyternotebook is created

* fix: incorrect yaml alignment

fix: changes following review

fix: correcting logging level

fix: ensure configmap is deleted when notebook is

fix: ensure configMap delete is graceful (doesn't error if it doesn't exist)

+++ code/workspaces/infrastructure-api/src/kubernetes/serviceAccountApi.js
@@ -1,11 +1,28 @@
 import logger from '../config/logger';
 import { getCoreV1Api } from './kubeConfig';
 
+async function namespacedServiceAccountExists(name, namespace) {
+  const k8sApi = getCoreV1Api();
+  logger.debug(`Checking serviceAccount ${name} exists in ${namespace}`);
+  try {
+    await k8sApi.readNamespacedServiceAccount(name, namespace);
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
 async function createNamespacedServiceAccount(name, namespace) {
+  if (await namespacedServiceAccountExists(name, namespace)) {
+    logger.debug(`ServiceAccount ${name} already exists in namespace ${namespace}`);
+    return;
+  }
+
   const k8sApi = getCoreV1Api();
 
   const serviceAccount = { metadata: { name } };
   try {
+    logger.info(`Creating ServiceAccount: ${name} in namespace ${namespace}`);
     await k8sApi.createNamespacedServiceAccount(namespace, serviceAccount);
   } catch (error) {
     logger.error(
@@ -16,9 +33,14 @@ async function createNamespacedServiceAccount(name, namespace) {
 }
 
 async function deleteNamespacedServiceAccount(name, namespace) {
+  if (!(await namespacedServiceAccountExists(name, namespace))) {
+    logger.debug(`ServiceAccount ${name} does not exist in namespace ${namespace}`);
+    return;
+  }
   const k8sApi = getCoreV1Api();
 
   try {
+    logger.info(`Deleting ServiceAccount: ${name} in namespace ${namespace}`);
     await k8sApi.deleteNamespacedServiceAccount(name, namespace);
   } catch (error) {
     logger.error(`Error deleting ServiceAccount ${name} in namespace ${namespace}: ${error.message}`);
